Fixing time dumping to SQL

+++ cmsat_tablestructure.sql
@@ -246,9 +246,9 @@ CREATE TABLE `timepassed` (
   `runID` bigint(20) unsigned NOT NULL,
   `simplifications` bigint(20) unsigned NOT NULL,
   `conflicts` bigint(20) unsigned NOT NULL,
-  `time` float unsigned NOT NULL,
+  `time` double unsigned NOT NULL,
   `name` varchar(200) NOT NULL,
-  `elapsed` float unsigned NOT NULL,
+  `elapsed` double unsigned NOT NULL,
   `timeout` int(20) unsigned NOT NULL,
   `percenttimeremain` float unsigned NOT NULL,
   KEY `idx1` (`runID`,`conflicts`)
